Split core logic into module
